requirements.txt
jinja2
pyinotify [AutoRenderer]
ruamel.yaml
setuptools
